fix(obs): Fix bug that in obs dynamic adjustment, the sum for obs value will exceed the limit of int16_t.

+++ src/movement.cpp
@@ -1611,7 +1611,7 @@ void obs_dynamic_base(uint16_t count)
 //	count = 20;
 //	enum {front,left,right};
 	static uint16_t obs_cnt[] = {0,0,0};
-	static int16_t obs_sum[] = {0,0,0};
+	static int32_t obs_sum[] = {0,0,0};
 	const int16_t obs_dynamic_limit = 2000;
 	int16_t* p_obs_baseline[] = {&g_obs_front_baseline, &g_obs_left_baseline, &g_obs_right_baseline};
 	typedef int16_t(*Func_t)(void);
